modify uplaod & delete response setting.

+++ app/backend/app/Services/Admins/BannersService.php
@@ -258,7 +258,7 @@ class BannersService
      * @param string $endAt end datetime
      * @param string $url url
      * @param UploadedFile|null $image image file
-     * @return \Illuminate\Http\JsonResponse
+     * @return void
      */
     public function updateBanner(
         string $uuid,
@@ -273,7 +273,7 @@ class BannersService
         string $endAt,
         string $url,
         ?UploadedFile $image
-    ): JsonResponse {
+    ): void {
         $resource = BannersResource::toArrayForUpdate(
             $uuid,
             $name,
@@ -293,6 +293,18 @@ class BannersService
             // ロックをかける為transaction内で実行
             $banner = $this->getBannerByUuid($uuid);
             $updatedRowCount = $this->bannersRepository->update($banner[Banners::ID], $resource);
+
+            // 更新出来ない場合
+            // 更新されていない場合は304を返すでも良さそう
+            if (!($updatedRowCount > 0)) {
+                throw new MyApplicationHttpException(
+                    StatusCodeMessages::STATUS_401,
+                    parameter: [
+                        'resource' => $resource,
+                        'banner.id' => $banner[Banners::ID],
+                    ]
+                );
+            }
 
             // 画像がアップロードされている場合
             if ($image) {
@@ -318,11 +330,7 @@ class BannersService
             // キャッシュの削除
             CacheLibrary::deleteCache(self::CACHE_KEY_ADMIN_BANNER_COLLECTION_LIST, true);
 
-            // 更新されていない場合は304
-            $message = ($updatedRowCount > 0) ? 'success' : 'not modified';
-            $status = ($updatedRowCount > 0) ? 200 : 304;
-
-            return response()->json(['message' => $message, 'status' => $status], $status);
+            // return response()->json(['message' => $message, 'status' => $status], $status);
         } catch (Exception $e) {
             Log::error(__CLASS__ . '::' . __FUNCTION__ . ' line:' . __LINE__ . ' ' . 'message: ' . json_encode($e->getMessage()));
             DB::rollback();
@@ -334,9 +342,9 @@ class BannersService
      * delete banner data service
      *
      * @param array<int, string> $bannerUuids
-     * @return \Illuminate\Http\JsonResponse
+     * @return void
      */
-    public function deleteBanner(array $bannerUuids): JsonResponse
+    public function deleteBanner(array $bannerUuids): void
     {
         DB::beginTransaction();
         try {
@@ -347,16 +355,24 @@ class BannersService
 
             $deleteRowCount = $this->bannersRepository->delete(array_column($banners, Banners::ID), $resource);
 
+            // 削除出来ない場合
+            // 更新されていない場合は304を返すでも良さそう
+            if (!($deleteRowCount > 0)) {
+                throw new MyApplicationHttpException(
+                    StatusCodeMessages::STATUS_401,
+                    parameter: [
+                        'resource' => $resource,
+                        'bannerUuids' => $bannerUuids,
+                    ]
+                );
+            }
+
             DB::commit();
 
             // キャッシュの削除
             CacheLibrary::deleteCache(self::CACHE_KEY_ADMIN_BANNER_COLLECTION_LIST, true);
 
-            // 更新されていない場合は304
-            $message = ($deleteRowCount > 0) ? 'success' : 'not deleted';
-            $status = ($deleteRowCount > 0) ? 200 : 401;
-
-            return response()->json(['message' => $message, 'status' => $status], $status);
+            // return response()->json(['message' => $message, 'status' => $status], $status);
         } catch (Exception $e) {
             Log::error(__CLASS__ . '::' . __FUNCTION__ . ' line:' . __LINE__ . ' ' . 'message: ' . json_encode($e->getMessage()));
             DB::rollback();
@@ -391,7 +407,12 @@ class BannersService
             if (!$result) {
                 throw new MyApplicationHttpException(
                     StatusCodeMessages::STATUS_500,
-                    'store file failed.'
+                    'store file failed.',
+                    [
+                        'fileResource' => $fileResource,
+                        'bannerId' => $bannerId,
+                        'storageFileName' => $storageFileName,
+                    ]
                 );
             }
